int-to-uint64

+++ pkg/security/secl/model/consts_linux_ppc64le.go
@@ -31,5 +31,5 @@ var (
                 "PTRACE_SYSEMU_SINGLESTEP":    unix.PTRACE_SYSEMU_SINGLESTEP,
         }
 
-        mmapFlagArchConstants = map[string]int{}
+        mmapFlagArchConstants = map[string]uint64{}
 )
